mcp/aws-mcp/requirements.txt
mcp>=1.0.0
boto3>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
